Put cross compile assertions in their own file
